Refactor socketio into seperate file
